fix(app): async first-run plist install (H6)
The first-run install path called origin-server install + launchctl load
synchronously in Tauri's setup() callback, blocking app start by
hundreds of ms (subprocess spawn + launchctl load are slow).

Extract the install flow into lifecycle::first_run_install_if_needed,
and run it on a spawn_blocking task wrapped in async_runtime::spawn.
On failure emit `origin-fallback-mode` so the UI can show a toast (the
sidecar fallback at the bottom of setup() still spawns the daemon).
